+++ concordium-std/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Change the `receive_name` parameter of `HasActions::send` to use `ReceiveName`
   instead of `str`.
 - Rename `send` to `send_raw` in `HasActions`.
+- Rename `log_bytes` to `log_raw` in `HasLogger`.
 - Add `send`, a wrapper for `HasActions::send_raw`, which automatically
   serializes `parameter` (using `Serial`).
 - Allow init and receive methods to return custom error codes that will be displayed to the user